feat: add skeleton codes for post, put, delete api handler
